requirements.txt
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
mne==1.4.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1
flask==2.2.5
flask-restful==0.3.10
joblib==1.3.1
PyWavelets==1.4.1
pytest==7.4.0
python-dotenv==1.0.0
